extract components, sort imports
