+++ src/components/CargoRemaining/index.css
@@ -1,0 +1,5 @@
+.cargo-remaining {
+  margin-top: 8px;
+  padding-top: 4px;
+  border-top: 1px solid #e5e5e5;
+}
