+++ src/components/CommodityIcon/index.css
@@ -1,0 +1,3 @@
+.commodity-icon {
+  cursor: default;
+}
