+++ src/components/ProjectLink/index.css
@@ -1,0 +1,8 @@
+.project-link .project-name {
+  font-weight: bold;
+}
+
+.grey {
+  background-color: rgb(238, 238, 238);
+  padding: 0 4px 0 4px;
+}
