+++ src/components/RecentProjects/index.css
@@ -1,4 +1,3 @@
-
 .recent {
   padding: 4px 0 0 4px;
   border-left: #e5e5e5 1px solid;
