--- src/misc.css
@@ -1,19 +1,0 @@
-
-.project-link .project-name {
-  font-weight: bold;
-}
-
-.grey {
-  background-color: rgb(238, 238, 238);
-  padding: 0 4px 0 4px;
-}
-
-.cargo-remaining {
-  margin-top: 8px;
-  padding-top: 4px;
-  border-top: 1px solid #e5e5e5;
-}
-
-.commodity-icon {
-  cursor: default;
-}
