+++ src/misc.tsx
@@ -1,72 +1,10 @@
-import './misc.css';
-import { FunctionComponent } from "react";
-import { mapCommodityIcon, mapCommodityNames, mapCommodityType, ProjectRefLite } from "./types";
-import { Icon } from "@fluentui/react";
-import { Store } from "./local-storage";
-import { buildTypes } from './project-create';
 import cargoTypes from './assets/cargo-types.json';
+import { mapCommodityNames } from "./types";
 
-interface ProjectLinkProps {
-  proj: ProjectRefLite;
-  noSys?: boolean;
-}
-
-export const ProjectLink: FunctionComponent<ProjectLinkProps> = (props) => {
-  // {!props.noSys && <><Icon iconName='LocationOutline' /> {props.proj.systemName}: </>}<a href={`#build=${props.proj.buildId}`}><Icon iconName='CityNext2' /> {props.proj.buildName}</a> ({props.proj.buildType})
-  return <span className="project-link">
-    {!props.noSys && <><a href={`#find=${props.proj.systemName}`}>{props.proj.systemName}</a> : </>}
-    <a className="project-name" href={`#build=${props.proj.buildId}`}><Icon iconName='Manufacturing' /> {props.proj.buildName}</a>
-    &nbsp;- <BuildType buildType={props.proj.buildType} />
-  </span>;
-};
-
-
-export const CommodityIcon: FunctionComponent<{ name: string }> = (props) => {
-
-  let commodityClass = getTypeForCargo(props.name);
-  let iconName = mapCommodityIcon[commodityClass]!;
-
-  if (commodityClass) {
-    iconName = mapCommodityIcon[commodityClass]!;
-  } else if (props.name in mapCommodityIcon) {
-    iconName = mapCommodityIcon[props.name];
-  } else {
-    console.error(`Unexpected: ${props.name}`);
-    commodityClass = 'Unknown';
-    iconName = 'ChromeClose';
-
-    mapCommodityType[props.name] = 'xxx'
-    console.log(mapCommodityType);
-  }
-
-  return <Icon className="commodity-icon" iconName={iconName} title={commodityClass} />;
-};
-
-
-export const CargoRemaining: FunctionComponent<{ sumTotal: number }> = (props) => {
-  const cmdr = Store.getCmdr();
-  const tripsLarge = Math.ceil(props.sumTotal / (cmdr?.largeMax ?? 794));
-  const tripsMed = Math.ceil(props.sumTotal / (cmdr?.medMax ?? 400));
-
-  return <div className="cargo-remaining hint">
-    Remaining cargo to deliver: <span className='grey'>{props.sumTotal.toLocaleString()}</span>
-    <br />
-    Large ship:<span className='grey'>{tripsLarge} trips</span> or Medium ship:<span className='grey'>{tripsMed} trips</span>
-  </div>;
-};
-
-export const BuildType: FunctionComponent<{ buildType: string }> = (props) => {
-
-  const match = buildTypes.find(i => i.text.toLowerCase().includes(props.buildType));
-
-  if (!match) {
-    console.error(`Why no match for: ${props.buildType} ?`)
-    return <span>{props.buildType}</span>;
-  }
-
-  // remove the trailing "(aa, bb, cc)" so we can use just the build type in question
-  const displayName = match.text.substring(0, match.text.indexOf('('));
-  return <span key={`bt${props.buildType}`}>{displayName}({props.buildType})</span>;
+export const getColorTable = (tokens: string[]): Record<string, string> => {
+  const colors: Record<string, string> = {};
+  tokens.forEach((k, n) => colors[k] = `qualitative.${n + 1}`);
+  return colors;
 };
 
 export const getTypeForCargo = (cargo: string) => {
