+++ src/views/Commander/index.css
@@ -18,7 +18,6 @@
   cursor: default;
 }
 
-
 .cmdr-projects .need {
   padding-right: 20px;
 }
