+++ src/views/Home.tsx
@@ -1,5 +1,5 @@
 import { PrimaryButton } from '@fluentui/react';
-import { RecentProjects } from '../recent-projects';
+import { RecentProjects } from '../components';
 
 export const Home: React.FunctionComponent = () => {
   window.document.title = `Raven Colonial Corporation`;
