+++ src/views/ProjectView/index.css
@@ -81,7 +81,6 @@
   display: inline;
 }
 
-
 .project-edit input {
   display: block;
   width: 400px;
@@ -97,7 +96,6 @@
   padding: 2px;
   white-space: pre-line;
 }
-
 
 .stats-header {
   margin-bottom: 10px;
